requirements.txt
streamlit
Pillow
scikit-learn
streamlit-flow-component
opencv-python
